requirements.txt
boto3>=1.26.0
pandas>=1.5.0
openpyxl>=3.0.10
python-dotenv>=1.0.0
pathlib2>=2.3.7
more-itertools>=8.0.0 